Refactor the classes so the code makes more sense
