+++ src/main/java/com/birdie/kafka/connect/json/SchemaTransformer.java
@@ -1,12 +1,12 @@
-package com.birdie.kafka.connect.smt;
+package com.birdie.kafka.connect.json;
 
+import com.birdie.kafka.connect.utils.StructWalker;
 import org.apache.kafka.connect.data.*;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
-import java.net.URISyntaxException;
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
